modification to include enum values into python
